fix: Creatures representation glitch
Removed the update to visual representations of the Creatures from the MainGameScene (excepted when the Scene is resized) as they are already updated by their own Timelines.

This caused a glitch when the competing Timelines updated the representations as the MainGameScene timeline could set it on its destination tile for a few milliseconds while the creature's own timeline was in between 2 keyframes.

updateGameView() renamed updateVisibilityAndFieldOfView() to better suit its restricted responsibility.